TinyML_robustness_analysis/requirements.txt
absl-py==2.1.0
adversarial-robustness-toolbox==1.18.2
asttokens==2.4.1
astunparse==1.6.3
boto3==1.35.42
botocore==1.35.42
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
filetype==1.2.0
fire==0.7.0
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.2
h5py==3.12.1
huggingface-hub==0.24.7
idna==3.7
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
keras==2.14.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.4.1
numpy==1.26.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.16.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybboxes==0.1.6
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
roboflow==1.1.48
rsa==4.9
s3transfer==0.10.3
sahi==0.11.18
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.14.0
tensorflow-addons==0.23.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
tensorrt==8.5.3.1
termcolor==2.5.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
torch==2.5.0+cu124
torchaudio==2.5.0+cu124
torchvision==0.20.0+cu124
tqdm==4.66.5
traitlets==5.14.3
triton==3.1.0
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.0.217
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.14.1
yolov5==7.0.13